+++ command_deck.h
@@ -2,18 +2,25 @@
 
 #include <furi.h>
 #include <gui/gui.h>
+#include <gui/modules/number_input.h>
 #include <gui/modules/submenu.h>
+#include <gui/modules/text_input.h>
+#include <gui/modules/variable_item_list.h>
 #include <gui/modules/widget.h>
 #include <gui/view_dispatcher.h>
 
 #include "command_deck_api.h"
 #include "command_deck_config.h"
+#include "command_deck_storage.h"
 
-#define COMMAND_DECK_VERSION "0.2.0"
+#define COMMAND_DECK_VERSION "0.3.0"
 
 typedef enum {
     CommandDeckViewMenu,
     CommandDeckViewStatus,
+    CommandDeckViewSettings,
+    CommandDeckViewHostInput,
+    CommandDeckViewPortInput,
 } CommandDeckView;
 
 typedef enum {
@@ -24,6 +31,7 @@ typedef enum {
     CommandDeckMenuUtilities,
     CommandDeckMenuCustomActions,
     CommandDeckMenuSystemStatus,
+    CommandDeckMenuSettings,
 } CommandDeckMenuItem;
 
 typedef struct {
@@ -31,5 +39,10 @@ typedef struct {
     ViewDispatcher* view_dispatcher;
     Submenu* menu;
     Widget* status;
+    VariableItemList* settings;
+    TextInput* host_input;
+    NumberInput* port_input;
     CommandDeckConfig config;
+    char host_edit[COMMAND_DECK_COMPANION_HOST_MAX];
+    char port_text[6];
 } CommandDeckApp;