+++ command_deck_storage.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "command_deck_config.h"
+
+bool command_deck_storage_load(CommandDeckConfig* config);
+bool command_deck_storage_save(const CommandDeckConfig* config);